+++ apps/frontend/src/app/modules/loader/services/remote-file-history/remote-file-history.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+
+import { RemoteFileHistoryService } from './remote-file-history.service';
+
+describe('RemoteFileHistoryService', () => {
+  let service: RemoteFileHistoryService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(RemoteFileHistoryService);
+  });
+
+  it('should be created', () => {
+    expect(service).toBeTruthy();
+  });
+});