+++ apps/frontend/src/app/modules/shared/public-api.ts
@@ -1,4 +1,5 @@
 export { IScenario } from './models/scenario';
 export { ScenariiRegistryService } from './services/scenarii-registry/scenarii-registry.service';
 export { ReportsRegistryService } from './services/reports-registry/reports-registry.service';
+export { DataManagerService } from './services/data-manager/data-manager.service';
 export { SharedModule } from './shared.module';