+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fix the remote file upload
+- Support remote URL history
 - Add some error messages on file upload
 
 ## [0.1.0] - 2021-12-22